% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_tools.R
\name{plot_km}
\alias{plot_km}
\title{Plot survival cures}
\usage{
plot_km(
  dat_surv,
  var_time,
  var_status,
  lab_y = "PFS",
  by_var = c("ARM"),
  event = 0,
  pval = FALSE,
  lim_x = 0,
  lim_y = c(0, 1)
)
}
\description{
Plot survival cures
}
