% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_design_present.R
\name{tb_des_plt_pfs_tbcut}
\alias{tb_des_plt_pfs_tbcut}
\title{Plot survival by response}
\usage{
tb_des_plt_pfs_tbcut(
  dta_tb,
  dta_surv,
  tb_cut = -0.3,
  var_tb = "PCHG",
  var_id = "SUBJID",
  var_time = "PFS_DAYS",
  var_status = "PFS_CNSR",
  ...
)
}
\description{
Plot survival by response
}
