% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_design_present.R
\name{tb_des_plt_tb_resp_waterf}
\alias{tb_des_plt_tb_resp_waterf}
\title{Waterfall plot of tumor burden}
\usage{
tb_des_plt_tb_resp_waterf(dat_sub, by_var = "ARM")
}
\description{
Waterfall plot of tumor burden
}
