% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_design_present.R
\name{tb_des_plt_tb_sub}
\alias{tb_des_plt_tb_sub}
\title{Spider plot of tumor burden}
\usage{
tb_des_plt_tb_sub(dat_sub, by_var = "ARM", tb_cut = -0.3, sample_n = NULL)
}
\description{
Spider plot of tumor burden
}
