% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_design.R
\name{tb_des_simu_all}
\alias{tb_des_simu_all}
\title{Simulate data for design}
\usage{
tb_des_simu_all(
  true_par,
  n = g_sample_n,
  enroll_dur_mth = 6,
  annual_dropout = 0.05,
  min_fu_mth = 12,
  visit_days = seq(9, 120, by = 9),
  visit_max_days = 300,
  miss_rate = g_tb_mis_rate,
  arms = c("Control", "Treatment"),
  seed = NULL
)
}
\description{
Simulate data for design
}
