% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_design.R
\name{tb_des_simu_pfs}
\alias{tb_des_simu_pfs}
\title{Simulate PFS in Days}
\usage{
tb_des_simu_pfs(
  par_surv,
  n = 500,
  sub_rand = NULL,
  label = NULL,
  mth_to_days = 1
)
}
\arguments{
\item{sub_rand}{Random characteristics of each subject}

\item{vec_c}{Time without 0}
}
\description{
Simulate PFS in Days
}
