% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_design.R
\name{tb_des_simu_tb}
\alias{tb_des_simu_tb}
\title{Simulate tumor burden for study design}
\usage{
tb_des_simu_tb(
  vec_t,
  par_tb,
  par_tb_rand = 0,
  par_tb_sig = 0.25,
  n = 500,
  sub_rand = NULL,
  label = "Control",
  rand_cut = 0,
  ...
)
}
\arguments{
\item{vec_t}{Time without 0}

\item{sub_rand}{Random characteristics of each subject}
}
\description{
Simulate tumor burden for study design
}
