% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_design.R
\name{tb_des_simu_tb_giv_surv}
\alias{tb_des_simu_tb_giv_surv}
\title{Simulate tumor burden for study design}
\usage{
tb_des_simu_tb_giv_surv(
  dta_surv,
  par_tb,
  par_tb_rand = c(0, 0, 0),
  par_tb_sig = 0.25,
  label = NULL
)
}
\arguments{
\item{sub_rand}{Random characteristics of each subject}

\item{vec_t}{Time without 0}
}
\description{
Simulate tumor burden for study design
}
