% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_design.R
\name{tb_des_surv_par}
\alias{tb_des_surv_par}
\title{Find survival related design parameters}
\usage{
tb_des_surv_par(
  median_surv,
  par_surv_rand = 0,
  label = "Control",
  nlarge = 20000,
  interval = c(-10, 10)
)
}
\arguments{
\item{median_surv}{Median survival in days}
}
\description{
Find survival related design parameters
}
