% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_design.R
\name{tb_des_surv_par_rand}
\alias{tb_des_surv_par_rand}
\title{Find survival related design parameters}
\usage{
tb_des_surv_par_rand(surv_days, surv_quants, label = "Control", nlarge = 20000)
}
\description{
Find survival related design parameters
}
