% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_design.R
\name{tb_des_tpar}
\alias{tb_des_tpar}
\title{Find tumor burden related design parameters}
\usage{
tb_des_tpar(
  mean_pchg,
  sd_pchg = NULL,
  vec_t = NULL,
  pred_vec_t = vec_t,
  poly_o = 3,
  label = "Control",
  ...
)
}
\description{
Find tumor burden related design parameters
}
