% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_design.R
\name{tb_des_tpar_rand}
\alias{tb_des_tpar_rand}
\title{Find tumor burden related design parameters}
\usage{
tb_des_tpar_rand(
  obs_surv,
  dta_surv,
  par_tb,
  par_tb_sig = 0.05,
  par_tb_rand = c(0, 0),
  k = 100,
  ...
)
}
\description{
Find tumor burden related design parameters
}
