% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_inference.R
\name{tb_estimate_summary}
\alias{tb_estimate_summary}
\title{Summarize estimation results}
\usage{
tb_estimate_summary(rst_estimate, arm_control = "Chemotherapy")
}
\description{
Summarize estimation results
}
