% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_tools_gamma.R
\name{tb_gamma_simu_bs}
\alias{tb_gamma_simu_bs}
\title{Evaluate treatment effect by AUC_SURV by BS}
\usage{
tb_gamma_simu_bs(
  simu_pt,
  uti_gamma,
  n_bs = 200,
  seed = 10000,
  n_cores = 5,
  ...
)
}
\description{
Evaluate treatment effect by AUC_SURV by BS
}
