% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_tools_gamma.R
\name{tb_gamma_simu_surv}
\alias{tb_gamma_simu_surv}
\title{Simulate survival outcomes for RCT}
\usage{
tb_gamma_simu_surv(n = 200, hd_death = c(1, 1.5), ...)
}
\description{
Simulate survival outcomes for RCT
}
