% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_inference.R
\name{tb_get_all_bs}
\alias{tb_get_all_bs}
\title{Overall results with bootstrap results}
\usage{
tb_get_all_bs(rst_orig, nbs = 100, seed = 1234, n_cores = 5)
}
\description{
Overall results with bootstrap results
}
