% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_utility.R
\name{tb_get_para_est}
\alias{tb_get_para_est}
\title{Extract Parameters}
\usage{
tb_get_para_est(
  surv_fit,
  tb_fit,
  trt_effect_surv = NULL,
  trt_effect_tb = NULL,
  tb_sig = NULL
)
}
\description{
Extract Parameters
}
