% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_patient.R
\name{tb_get_pt}
\alias{tb_get_pt}
\title{Get patient data}
\usage{
tb_get_pt(
  id,
  imp_surv,
  dat_tb,
  date_dbl = NULL,
  t_ana = NULL,
  imp_inx = 1,
  reg_tb = NULL,
  by_arm = TRUE,
  ...
)
}
\arguments{
\item{date_dbl}{Date of database lock}
}
\description{
Get turmor burden, survival and utility data of a specific patient
}
