% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_surv_msm.R
\name{tb_msm_imp_single}
\alias{tb_msm_imp_single}
\title{Impute survival for single patient}
\usage{
tb_msm_imp_single(d, fit_rst, imp_m, ind_event = 0, ...)
}
\description{
Impute survival for single patient
}
