% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_surv.R
\name{tb_msm_imp_surv}
\alias{tb_msm_imp_surv}
\title{Get survival data for multi-state model}
\usage{
tb_msm_imp_surv(msm_surv, formula_surv, ..., seed = 10000)
}
\description{
Get survival data for multi-state model
}
