% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_present.R
\name{tb_plt_ind}
\alias{tb_plt_ind}
\title{Plot a patient}
\usage{
tb_plt_ind(
  pt_his,
  ylim = NULL,
  xlim = NULL,
  type = c("uti", "under"),
  add_reflines_x = TRUE,
  add_reflines_y = TRUE,
  add_poly = TRUE,
  add_title = TRUE,
  add_line = TRUE,
  add_tb_obs = TRUE
)
}
\description{
Plot a patient
}
