% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_present.R
\name{tb_plt_km_imp}
\alias{tb_plt_km_imp}
\title{Survival curves for imputed survival}
\usage{
tb_plt_km_imp(imp_surv, dat_surv, inx_imp = NULL, type = c("PFS", "OS"), ...)
}
\description{
Survival curves for imputed survival
}
