% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_surv.R
\name{tb_plt_onstudy}
\alias{tb_plt_onstudy}
\title{Plot patients by enrollment}
\usage{
tb_plt_onstudy(
  t_enroll,
  t_time,
  event,
  t_dur,
  add_auc = FALSE,
  auc_k = 1.2,
  add_lab = TRUE,
  size_lab = 8,
  hjust_lab = -1,
  h = 0.4
)
}
\description{
Plot patients by enrollment
}
