% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_present.R
\name{tb_plt_surv}
\alias{tb_plt_surv}
\title{Plot survival curve with area under the curve}
\usage{
tb_plt_surv(
  surv_f,
  t_dur = NULL,
  type = c("rmf", "rmst"),
  ylim = NULL,
  xlim = NULL
)
}
\description{
Plot survival curve with area under the curve
}
