% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_tbreg.R
\name{tb_prepare_tb_dta}
\alias{tb_prepare_tb_dta}
\title{Combine survival and tumor burden data}
\usage{
tb_prepare_tb_dta(
  dat_tb,
  imp_surv,
  imp_inx = 1,
  uti_gamma = c(Progression = 0.2, Death = 0.5),
  tb_reg_pow_conv = 1,
  tb_reg_add_surv = FALSE,
  ...
)
}
\arguments{
\item{add_surv}{Whether include uti_gamma at the event time as a tumor burden
measurement}

\item{f_conv}{Function for converting tumor burden}
}
\description{
Prepare data set for longitudinal tumor burden regression
}
