% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_tb_tools.R
\name{tb_pseu_or}
\alias{tb_pseu_or}
\title{Define Pseudo Overall Response}
\usage{
tb_pseu_or(pchg, vec_t = NULL, thresh_cr = -0.3, thresh_pd = 0.2)
}
\arguments{
\item{thresh_cr}{Threshold for CR/PR: >=30% shrinkage from baseline}

\item{thresh_pd}{Threshold for PD: >= 20% increase than the minimum}
}
\description{
Define Pseudo Overall Response
}
