% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_patient_curve.R
\name{tb_pt_curve_emp_single}
\alias{tb_pt_curve_emp_single}
\title{TB linear prediction points}
\usage{
tb_pt_curve_emp_single(tb_mat, ts)
}
\description{
TB prediction based on the observed the utility curve
}
