% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_patient_curve.R
\name{tb_pt_curve_summary}
\alias{tb_pt_curve_summary}
\title{TB curves summarization}
\usage{
tb_pt_curve_summary(dta_curve, quant = NULL)
}
\description{
Summarize TB curves with Jackknife variances
}
