% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_patient.R
\name{tb_pt_insert}
\alias{tb_pt_insert}
\title{Insert tb and time}
\usage{
tb_pt_insert(
  tb_mat,
  ins_t,
  ins_ind,
  uti = 0.2,
  event_window = 7,
  method = c("composite", "locf", "extrap"),
  ...
)
}
\description{
Insert tb and time
}
