% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_tb_reg.R
\name{tb_regression}
\alias{tb_regression}
\title{Tumor burden regression}
\usage{
tb_regression(dat_tb, imp_surv, scale_time_by_surv = TRUE, by_arm = TRUE, ...)
}
\description{
Conduct tumor burden regression
}
