% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_tbreg.R
\name{tb_regression_single}
\alias{tb_regression_single}
\title{Tumor burden regression}
\usage{
tb_regression_single(
  tb_surv,
  fml_tb = NULL,
  tb_reg_poly_order = 2,
  tb_reg_pow_conv = 1,
  ...
)
}
\description{
Conduct tumor burden regression
}
