% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_tools_shiny.R
\name{tb_shiny}
\alias{tb_shiny}
\title{Run Web-Based application}
\usage{
tb_shiny(appname = "shiny")
}
\description{
Call Shiny to run \code{ctdesign} as a web-based application.
}
\details{
A web browser will be brought up for users to access the GUI
}
