% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_simu.R
\name{tb_simu_all}
\alias{tb_simu_all}
\title{Simulate studies}
\usage{
tb_simu_all(
  dat_tb,
  n,
  surv_fit,
  tb_fit,
  trt_effect_surv = 0,
  trt_effect_tb = 0,
  rand_effect_surv = 0,
  rand_effect_tb = c(`(Intercept)` = 0, `poly(reg_t, 2, raw = TRUE)1` = 0, `poly(reg_t,
    2, raw = TRUE)2` = 0),
  ...,
  seed = NULL
)
}
\description{
Simulate studies
}
