% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_simu_cov.R
\name{tb_simu_cov_es}
\alias{tb_simu_cov_es}
\title{Simulate covariates}
\usage{
tb_simu_cov_es(
  dta_es,
  v_covs,
  n = 500,
  label_arm = "ARM",
  label_id = "SUBJID",
  label_randt = "RANDT",
  rnd_ratio_trt = 1,
  replace = TRUE,
  ...,
  seed = NULL
)
}
\arguments{
\item{dta_es}{Existing study data}

\item{v_covs}{Vector of covariates to be sampled}

\item{n}{Sample size for two arms}

\item{label_arm}{Column name for Arm in the result}

\item{label_id}{Column name for patient id in the existing data}

\item{rnd_ratio_trt}{Randomization ratio vs. control (=1)}

\item{replace}{Sample with or without replacement}
}
\description{
Simulate covariates based on existing real studies
}
