% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_simu_surv.R
\name{tb_simu_surv}
\alias{tb_simu_surv}
\title{Simulate Survival time for all}
\usage{
tb_simu_surv(
  n,
  hd_prog = 0.5,
  hd_death = 1,
  hd_cens = 1,
  dur_enroll = 1,
  dur_study = 3,
  seed = NULL
)
}
\description{
Simulate Survival time for all
}
