% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_simu_surv.R
\name{tb_simu_surv_wb}
\alias{tb_simu_surv_wb}
\title{Simulate Survival time for all based on weibull}
\usage{
tb_simu_surv_wb(
  pt_cov,
  mdl_fit,
  trt_effect = 0,
  rand_effect = 0.1,
  annual_dropout = 0.1,
  min_fu_days = 0.6 * 365.25,
  ...,
  seed = NULL
)
}
\arguments{
\item{min_fu_day}{minimum follow up days}
}
\description{
Simulate Survival time for all based on weibull
}
