% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_present.R
\name{tb_summary_imp}
\alias{tb_summary_imp}
\title{Summarize imputed survival data}
\usage{
tb_summary_imp(imp_surv, dat_surv, inx_imp = NULL, by_var = c("ARM"))
}
\description{
Summarize imputed survival data
}
