% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_tb_tools.R
\name{tb_tb_impr}
\alias{tb_tb_impr}
\title{Get improved tumor burden}
\usage{
tb_tb_impr(vec, perc_improve = 0.1, type = c("ratio", "absolute"))
}
\arguments{
\item{perc_improve}{Percentage improvement}
}
\description{
Get improved tumor burden
}
