% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_tb_tools.R
\name{tb_tb_obs}
\alias{tb_tb_obs}
\title{Observed TB mean, sd and missing rate}
\usage{
tb_tb_obs(
  dta_tb,
  dta_surv,
  var_tb = "PCHG",
  var_day = "DAY",
  var_id = "SUBJID",
  var_time = "PFS_DAYS"
)
}
\description{
@export
}
