% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_tb_tools.R
\name{tb_tb_surv}
\alias{tb_tb_surv}
\title{Median survival by best TB}
\usage{
tb_tb_surv(
  dta_tb,
  dta_surv,
  tb_cut = -0.3,
  var_tb = "PCHG",
  var_id = "SUBJID",
  var_time = "PFS_DAYS",
  var_status = "PFS_CNSR",
  surv_quants = c(0.5, 0.7),
  ...
)
}
\arguments{
\item{cut_tb}{Best response measured by PCHG

 @export}
}
\description{
Median survival by best TB
}
