% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_utility.R
\name{tb_uti_surv_obs_single}
\alias{tb_uti_surv_obs_single}
\title{Utility based on observed survival only}
\usage{
tb_uti_surv_obs_single(t_event, t_dur, uti_gamma, type = c("jump", "connect"))
}
\description{
Utility based on observed survival only
}
