% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_utility.R
\name{tb_uti_surv_prob_single}
\alias{tb_uti_surv_prob_single}
\title{Utility based on survival function only}
\usage{
tb_uti_surv_prob_single(surv_f, t_dur, uti_gamma, type = c("jump", "connect"))
}
\description{
Utility based on survival function only
}
