% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_surv_weibull.R
\name{tb_weibull_fit}
\alias{tb_weibull_fit}
\title{Fit PFS Weibull}
\usage{
tb_weibull_fit(
  dat_surv,
  fml_surv,
  type_surv = c("PFS", "OS"),
  ind_event = 0,
  by_arm = TRUE,
  ...
)
}
\description{
Fit PFS Weibull
}
