% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_surv_weibull.R
\name{tb_weibull_imp_single}
\alias{tb_weibull_imp_single}
\title{Impute survival for single patient based on weibull regression}
\usage{
tb_weibull_imp_single(d, fit_rst, imp_m, ...)
}
\description{
Impute survival for single patient based on weibull regression
}
