% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkits.R
\name{tkt_simu_enroll}
\alias{tkt_simu_enroll}
\title{Simulate Enrollment Time}
\usage{
tkt_simu_enroll(
  ntot,
  enroll_duration,
  min_fu = 6,
  date_bos = NULL,
  mth_to_days = 30.4
)
}
\description{
Simulate Enrollment Time
}
