% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkits.R
\name{tkt_survfit_median}
\alias{tkt_survfit_median}
\title{Get table of median from survfit results}
\usage{
tkt_survfit_median(fit, surv_quants = NULL)
}
\description{
Get table of median from survfit results
}
