% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_inference.R
\name{tb_get_all}
\alias{tb_get_all}
\title{Overall results}
\usage{
tb_get_all(
  dat_tb,
  dat_surv,
  inx_bs = 0,
  formula_surv = "Surv(time,status)~trans+ARM+AGE+SEX+STRATA1+P1TERTL",
  imp_m = 5,
  date_dbl = "2020-03-01",
  gamma = c(0.2, 0.5)
)
}
\description{
Overall results
}
