% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_surv.R
\name{tb_plt_surv}
\alias{tb_plt_surv}
\title{Plot survival curve with area under the curve}
\usage{
tb_plt_surv(
  surv_f,
  t_dur = NULL,
  type = c("rmf", "rmst", "none"),
  y_lim = c(0, 1),
  x_lim = NULL
)
}
\description{
Plot survival curve with area under the curve
}
