% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp_patient.R
\name{tb_pt_hist}
\alias{tb_pt_hist}
\title{Generate history of a patient}
\usage{
tb_pt_hist(time_tb, tb, time_event, t_ana, gamma = c(0.2, 0.5), locf = FALSE)
}
\description{
Generate history of a patient
}
